refactor(workstation): extract yank + changelog actions into hooks (0.72 app.ts decomposition)
